+++ service/config.go
@@ -5,6 +5,7 @@ import "time"
 // Service represents particular service configuration
 type Service struct {
 	Command         string        `mapstructure:"command"`
+	Output          string        `mapstructure:"log_output"`
 	ProcessNum      int           `mapstructure:"process_num"`
 	ExecTimeout     time.Duration `mapstructure:"exec_timeout"`
 	RemainAfterExit bool          `mapstructure:"remain_after_exit"`