+++ service/plugin.go
@@ -55,11 +55,11 @@ func (service *Plugin) Serve() chan error {
 			for i := 0; i < service.cfg.Services[k].ProcessNum; i++ {
 				// create processor structure, which will process all the services
 				service.processes = append(service.processes, NewServiceProcess(
+					service.cfg.Services[k].Output,
 					service.cfg.Services[k].RemainAfterExit,
 					service.cfg.Services[k].ExecTimeout,
 					service.cfg.Services[k].RestartSec,
 					service.cfg.Services[k].Command,
-					service.logger,
 					errCh,
 				))
 			}